# docs/api-reference/schemas/openapi-data-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Data API",
    "version": "0.1.0",
    "description": "Preview specification for the Data API."
  },
  "servers": [
    {
      "url": "https://data-api.kuest.com",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/holders": {
      "get": {
        "summary": "List top token holders",
        "description": "Ranks wallets by outcome token balance for each requested market.",
        "tags": [
          "Market data"
        ],
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "description": "Comma-separated list of condition IDs.",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Hash64"
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of holders to return per market (max 500).",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 500,
              "default": 100
            }
          },
          {
            "name": "minBalance",
            "in": "query",
            "description": "Drop wallets below this token balance.",
            "schema": {
              "type": "number",
              "minimum": 0,
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Holder leaderboard returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetaHolder"
                  }
                },
                "examples": {
                  "sample": {
                    "value": [
                      {
                        "token": "0xdd22472e552920b8438158ea7238bfadfa4f736aa4cee91a6b86c39ead110917:0",
                        "holders": [
                          {
                            "depositWallet": "0x56687bF447db6fFa42FfE2204a05edAa20f55839",
                            "bio": "Macro strategist backing major economic events.",
                            "asset": "0xdd22472e552920b8438158ea7238bfadfa4f736aa4cee91a6b86c39ead110917:0",
                            "pseudonym": "macro.whale",
                            "amount": 12500,
                            "displayUsernamePublic": true,
                            "outcomeIndex": 0,
                            "name": "Macro Whale",
                            "profileImage": "https://cdn.domain.com/users/macro-whale.png",
                            "profileImageOptimized": "https://cdn.domain.com/users/macro-whale@2x.jpg"
                          },
                          {
                            "depositWallet": "0x94d37B9c164888FbA4d4b3D4639a65f8AC54a1B6",
                            "bio": "On-chain analyst seeding liquidity.",
                            "asset": "0xdd22472e552920b8438158ea7238bfadfa4f736aa4cee91a6b86c39ead110917:0",
                            "pseudonym": "liquidity.lab",
                            "amount": 8200.75,
                            "displayUsernamePublic": false,
                            "outcomeIndex": 0,
                            "name": "Liquidity Lab",
                            "profileImage": "https://cdn.domain.com/users/liquidity-lab.png",
                            "profileImageOptimized": "https://cdn.domain.com/users/liquidity-lab@2x.jpg"
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "sample": {
                    "value": {
                      "error": "unknown_field: minBalance"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "sample": {
                    "value": {
                      "error": "unauthorized: ip_not_allowlisted"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "sample": {
                    "value": {
                      "error": "internal_error: unexpected database failure"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Address": {
        "type": "string",
        "description": "0x-prefixed EVM address (40 hex chars).",
        "pattern": "^0x[a-fA-F0-9]{40}$",
        "example": "0x56687bf447db6ffa42ffe2204a05edaa20f55839"
      },
      "Hash64": {
        "type": "string",
        "description": "0x-prefixed 64-character hex string.",
        "pattern": "^0x[a-fA-F0-9]{64}$",
        "example": "0xdd22472e552920b8438158ea7238bfadfa4f736aa4cee91a6b86c39ead110917"
      },
...
      "MetaHolder": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "holders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Holder"
            }
          }
        }
      },
...
      "Holder": {
        "type": "object",
        "properties": {
          "depositWallet": {
            "$ref": "#/components/schemas/Address"
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "asset": {
            "type": "string"
          },
          "pseudonym": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number"
          },
          "displayUsernamePublic": {
            "type": "boolean"
          },
          "outcomeIndex": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "profileImage": {
            "type": "string",
            "nullable": true
          },
          "profileImageOptimized": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}